Pictures
Here are the images for our program. I am having a little trouble creating some of our objects as icons only as they include background. so i changed some of them to match the background of the road (egg, cars etc). I also included two options for the world background. the 6 lanes may be a little difficult to work with in our coding so i also included one with 4 lanes. let me know your thoughts