Moved timer to road class
It turns out it's much easier to use the timer if it's part of the road class instead of an actor subclass. So I moved most of it into the road class, and deleted the original timer class. Also the text on the screen is now printed with the much simpler 'showText' function.