Changed the size of images
I made the car images smaller so that they would fit in the lanes easier. I also expanded the road's backgroujnd image to show the bottom green as the starting area. i changed the location of the score board and the timer to align with the expansion of the road background image.