Added functional start screen
On initialization, the game displays a start screen and a notice to start the game by using the space bar. Hitting the space bar sets the game in motion. Note the game must still have the "run" button clicked inside Greenfoot IDE, before pressing the space bar, for the game to run.